+++ next.config.ts
@@ -1,8 +1,0 @@
-import type { NextConfig } from 'next';
-
-const nextConfig: NextConfig = {
-  output: 'standalone',
-  allowedDevOrigins: ['10.0.0.75'],
-};
-
-export default nextConfig;